Screenshot Service: separate GLScreenshotSource file
